+++ test/corpus/fn-icon-libraries/expected.fixed.tsx
@@ -1,0 +1,11 @@
+import { BsChevronLeft } from "react-icons/bs";
+import ArrowLeftIcon from "./ArrowLeftIcon";
+import { MdKeyboardArrowRight } from "react-icons/md";
+import { ChevronsLeft, ChevronDown, Menu } from "lucide-react";
+
+export const Bar = () => (
+  <nav>
+    <ChevronDown />
+    <Menu />
+  </nav>
+);