+++ test/corpus/tailwind-happy/input.tsx
@@ -1,0 +1,5 @@
+export const Box = () => (
+  <div className="ml-4 mr-2 pl-1 pr-3 text-left float-right rounded-l-lg border-r-2 scroll-pl-4 left-0">
+    <span className="-ml-2 md:pr-6 hover:text-right !pl-2">x</span>
+  </div>
+);